pyproject.toml
# https://setuptools.pypa.io/en/latest/userguide/quickstart.html
[project]
name = "f"
version = "0.0.1"
dependencies = [
    # starfarm dependencies
    "ghapi",
    "gitpython",
    "PyYAML",
    "emoji",

    # note dependencies
    "python-frontmatter", # https://pypi.org/project/python-frontmatter/
    
    # note - extracting link feature
    "promnesia",
    "mistletoe",

    # bookmark dependencies
    "raindrop-io-py" # https://github.com/PBorocz/raindrop-io-py


]

[tool.setuptools.packages.find]
# All the following settings are optional:
where = ["f"]  # ["."] by default